chore: fix all lint errors and warnings
Bring `npm run lint` to 0 errors / 0 warnings without changing runtime
behavior, unblocking adding lint to CI.

- no-useless-escape (GraphContext.tsx, useDataset.ts): removed the
  unnecessary `\/` inside the `[^/]` character class in the path regex.
  The leading `\/` (which must stay) is unchanged; regex meaning identical.
- react-hooks/rules-of-hooks (CrossSceneTeaser.tsx): moved the useCallback
  above the early `return null` so hooks run in a stable order.
- react-refresh/only-export-components: moved non-component exports into
  sibling modules and updated all import sites:
    - DatasetTile.tsx -> DatasetTile.utils.ts (getStartYear)
    - ResourceMeta.tsx -> ResourceMeta.utils.ts (build*OgImageUrl + base URL)
    - Toast.tsx -> useToast.ts (useToast hook + ToastMessage type)
    - CrossSceneContext.tsx -> crossSceneContextCore.ts (context object,
      value/data types, useCrossSceneData/useCrossSceneDataOptional hooks)
- react-hooks/exhaustive-deps (Header.tsx): wrapped the setSearchTerm noop
  fallback in useMemo so its identity is stable (the fix the rule suggests).
- react-hooks/exhaustive-deps (ForceGraphLayout.tsx, RadialLayout.tsx):
  targeted eslint-disable with justification. getCrossSceneData changes
  identity as the cross-scene cache populates; adding it would force a full
  d3 graph re-layout on every cache update (behavior change / layout thrash).
