+++ src/components/index.ts
@@ -1,7 +1,7 @@
 export { default as Header } from './Header'
 export { default as MainLayout } from './MainLayout'
 export { default as DatasetTile } from './DatasetTile'
-export { getStartYear } from './DatasetTile'
+export { getStartYear } from './DatasetTile.utils'
 export { default as DatasetExploreWrapper } from './DatasetExploreWrapper'
 export { default as InfoboxPanel } from './InfoboxPanel'
 export { default as NodeInfobox } from './NodeInfobox'
@@ -28,5 +28,6 @@ export { default as WikipediaAttribution } from './WikipediaAttribution'
 export { default as FeedbackButton } from './FeedbackButton'
 export { default as FeedbackForm } from './FeedbackForm'
 export { default as GraphLegend } from './GraphLegend'
-export { default as Toast, useToast } from './Toast'
-export type { ToastMessage } from './Toast'
+export { default as Toast } from './Toast'
+export { useToast } from './useToast'
+export type { ToastMessage } from './useToast'
