+++ src/components/useToast.ts
@@ -1,0 +1,32 @@
+/**
+ * Toast state management.
+ *
+ * Kept in a separate module (not Toast.tsx) so the component file only exports
+ * components, which keeps React Fast Refresh working.
+ */
+
+import { useState, useCallback } from 'react';
+
+export interface ToastMessage {
+  id: string;
+  message: string;
+  type?: 'success' | 'info';
+}
+
+/**
+ * Hook for managing toast state
+ */
+export function useToast() {
+  const [toasts, setToasts] = useState<ToastMessage[]>([]);
+
+  const showToast = useCallback((message: string, type: 'success' | 'info' = 'success') => {
+    const id = `toast-${Date.now()}-${Math.random().toString(36).slice(2, 9)}`;
+    setToasts((prev) => [...prev, { id, message, type }]);
+  }, []);
+
+  const dismissToast = useCallback((id: string) => {
+    setToasts((prev) => prev.filter((t) => t.id !== id));
+  }, []);
+
+  return { toasts, showToast, dismissToast };
+}
