+++ src/contexts/index.ts
@@ -4,10 +4,10 @@
 
 export { GraphProvider, useGraph, useGraphOptional } from './GraphContext';
 export { ThemeProvider, useTheme } from './ThemeContext';
+export { CrossSceneProvider } from './CrossSceneContext';
 export {
-  CrossSceneProvider,
   useCrossSceneData,
   useCrossSceneDataOptional,
   type CrossSceneAppearance,
   type CrossSceneData,
-} from './CrossSceneContext';
+} from './crossSceneContextCore';
