+++ src/layouts/RadialLayout.tsx
@@ -336,6 +336,10 @@ export function RadialLayout({
 
     // No cleanup needed
     return () => {};
+    // getCrossSceneData is intentionally omitted: its identity changes whenever the
+    // cross-scene cache populates, and including it would force a full graph re-layout
+    // on every cache update. Cross-scene indicators are read as a best-effort snapshot.
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [data, dimensions, centerNode, connectedNodes, edges, hasValidSelection, onNodeClick, onEdgeClick]);
 
   // Update selection highlighting
