+++ src/pages/NodeDetailPage.tsx
@@ -16,7 +16,8 @@ import { sanitizeUrl, isValidImageUrl, getNodeTypeEmoji } from '@utils';
 import { useResourceParams, buildFullNodeUrl, buildGraphViewUrl } from '@hooks/useResourceParams';
 import { useNodeEnrichedData } from '@hooks';
 import { useCrossSceneAppearances } from '@hooks/useCrossSceneAppearances';
-import ResourceMeta, { buildNodeOgImageUrl } from '@components/ResourceMeta';
+import ResourceMeta from '@components/ResourceMeta';
+import { buildNodeOgImageUrl } from '@components/ResourceMeta.utils';
 import SchemaOrg from '@components/SchemaOrg';
 import WikipediaAttribution from '@components/WikipediaAttribution';
 import { CrossSceneSection } from '@components/CrossSceneSection';
